Complete SearchService

+++ SearchService.Domain/SearchBlogResponse.cs
@@ -1,5 +1,5 @@
 ﻿namespace SearchService.Domain
 {
     //搜索返回结果
-    public record SearchBlogResponse(IEnumerable<BlogRecord> Blogs,int TotalCount);
+    public record SearchBlogResponse(IEnumerable<BlogRecord> Blogs,long TotalCount);
 }
